Update predicate field RequestBody to have object type
